add likes test in VideoModelTest

+++ youtube/models.py
@@ -29,7 +29,7 @@ class Video(models.Model):
     slug = models.SlugField(unique=True, editable=False)
     uploaded_at = models.DateTimeField(auto_now_add=True)
     external_url = models.URLField(blank=True, null=True)
-    likes = models.IntegerField(verbose_name="likes", default=0)
+    likes = models.PositiveBigIntegerField(verbose_name="likes", default=0)
 
     def __str__(self):
         return self.title
